Leverage custom errors
Use custom errors that implement std::error::Error and
have them return nice messages formatted lower case with
no periods.

Note: If we need to provide more context around the errors
for debugging, I suggest we use the debug!/error! macros
instead of returning that in the exception string. I want
to keep the exception string fairly user friendly.

Also this removes pretty much all of the excepts and unwraps
that should actually fail. The goal there is to make the
code a bit more robust for things like calling it as a service.
If the code that leverages it needs to fail early we might
look into having it look at the error output from the logs
and make decisions based on that.