Made gitleaks scan work on windows
Windows does not have stdout, which it was using.

Output to a file `gitleaks-results.json` in the scan dir. This made the
most sense as it is removed after scanning any way and very very
unlikely to have a name clash. If it does clash it is written after the
scan is performed and will be overwritten with valid scan result data.