Log, dont fail, when a URL fetch fails in JSONData
The change is simple. It logs that a URL was unable to be fetched but
allows the rest of the JSONData scan, and URL fetching to continue.

This is a short term solution until we improve our general error
handling. Allowing fatal/nonfatal errors to be floated up to the
application.

Commented out the CloneBrokenURLWithFetchURLs test as a broken URL is
not a fatal error. This should be reinstated when better error handling
is implemented.

Added a few items to the todo list, including the afformentioned error
handling.