Log, dont fail, when a URL fetch fails in JSONData (#50)
The change is small, instead of passing the error up, which would cause the clone to fail and the JSONData to not be scanned. This logs that a URL was unable to be fetched, passing back no error and allowing scanning to continue.

This is a short term solution until we improve our general error handling. Allowing fatal/nonfatal errors to be floated up to the application.

Commented out the CloneBrokenURLWithFetchURLs test as a broken URL is
not a fatal error. This should be reinstated when better error handling
is implemented.

Added a few items to the to do list, including the aforementioned error
handling.