Don't create config/cache dirs until needed
This fixes an issue where you can get messages like:

```
could not create dir: path="/.cache/leaktk"
```

When you run leaktk-scanner as a subprocess, are passing in a specific
workdir from the config and HOME isn't set. Before it would try to
create ~/.cache/leaktk first, even if that's not where the cache dir
was configured to be.

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED