Refactor Git.IsLocal()
Check the presense of files at the resource location if local isn't
provided in the opts for a git repo.

This simplifies the code at the small risk of someone having a local
folder with the same name of a git ssh clone url user@host. But the risk
can be mitigated for cases like that by setting '{"local": false}'
when run run the manual scan command.

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED