Add docker/setup-qemu-action step
From the [docs](https://goreleaser.com/cookbooks/multi-platform-docker-images/)

For buildx to work properly, you'll need to install qemu. On GitHub
actions, the easiest way is to use docker/setup-qemu-action.  It's also
important that the FROM in your Dockerfile is multi-platform, otherwise
it'll not work.  As long as you have Qemu and Docker set up, everything
should just work.